+++ src/main/java/com/zpig333/runesofwizardry/tileentity/TileEntityDustDye.java
@@ -1,23 +1,21 @@
 package com.zpig333.runesofwizardry.tileentity;
 
-import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.inventory.IInventory;
 import net.minecraft.item.ItemStack;
 import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.nbt.NBTTagList;
 import net.minecraft.tileentity.TileEntity;
-import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TextComponentString;
-import net.minecraft.util.text.TextComponentTranslation;
+import net.minecraft.util.EnumFacing;
+import net.minecraftforge.common.capabilities.Capability;
+import net.minecraftforge.items.CapabilityItemHandler;
+import net.minecraftforge.items.ItemStackHandler;
 
-import com.zpig333.runesofwizardry.core.References;
 import com.zpig333.runesofwizardry.core.WizardryLogger;
-import com.zpig333.runesofwizardry.item.dust.DustDyed;
+import com.zpig333.runesofwizardry.core.WizardryRegistry;
+import com.zpig333.runesofwizardry.item.ItemDustPouch;
 
 
-public class TileEntityDustDye extends TileEntity implements IInventory{
+public class TileEntityDustDye extends TileEntity{
 	//only 1 slot for now, might change if dyes are required as input
-	private ItemStack[] contents = new ItemStack[1];
+	private DyedDustItemStackHandler inventory = new DyedDustItemStackHandler(this);
 	//the currently selected color
 	private String colorString;
 	//was this block powered last time we checked
@@ -29,27 +27,54 @@ public class TileEntityDustDye extends TileEntity implements IInventory{
 		//colorString=I18n.translateToLocal(References.Lang.COLOR);
 	}
 	public void dye(int color){
-		if(contents[0]==null)return;
+		//technically, this stack should not be modified
+		ItemStack dust = inventory.getStackInSlot(0);
+		if(dust==null)return;
+		ItemStack pouch = null;
+		ItemDustPouch itemPouch = null;
+		if(dust.getItem() instanceof ItemDustPouch){
+			pouch = dust;
+			itemPouch = (ItemDustPouch)pouch.getItem();
+			dust = itemPouch.getDustStack(pouch, 0);
+			if(dust==null || dust.getItem()!=WizardryRegistry.dust_dyed){
+				WizardryLogger.logError("the TEDustDye at "+getPos()+" had a pouch with null/non dyed dust");
+				return;
+			}
+		}
 		//NPE was because tagCompound is null...
-		NBTTagCompound compound = contents[0].getTagCompound();
+		NBTTagCompound compound = dust.getTagCompound();
 		if(compound == null){
 			compound = new NBTTagCompound();
-			contents[0].setTagCompound(compound);
+			dust.setTagCompound(compound);
 		}
 		compound.setInteger("color", color);
+		if(pouch!=null){
+			dust.stackSize=itemPouch.getDustAmount(pouch);
+			itemPouch.clear(pouch);
+			itemPouch.addDust(pouch, dust);
+		}
 		setColor(Integer.toHexString(color));
+		this.markDirty();
 	}
 	
 	public void handleBlockUpdate(boolean newRedstone){
-		if(newRedstone && !pastRedstoneState){
+		if(pastRedstoneState && !newRedstone){//falling edge
 			try{
 				this.dye(Integer.parseInt(colorString,16));
 			}catch(NumberFormatException e){
 				WizardryLogger.logInfo("Dust Dye: unable to parse color "+colorString+" on redstone toggle");
 			}
 		}
+		pastRedstoneState=newRedstone;
 	}
 	
+	/* (non-Javadoc)
+	 * @see net.minecraft.tileentity.TileEntity#getUpdateTag()
+	 */
+	@Override
+	public NBTTagCompound getUpdateTag() {
+		return writeToNBT(super.getUpdateTag());
+	}
 	/**
 	 * 
 	 * @return the currently selected Color of this block as a String
@@ -64,128 +89,32 @@ public class TileEntityDustDye extends TileEntity implements IInventory{
 	public void setColor(String color){
 		colorString=color;
 	}
-	@Override
-	public int getSizeInventory() {
-		return contents.length;
-	}
+	
 
-	@Override
-	public ItemStack getStackInSlot(int i1) {
-		return contents[i1];
-	}
-
-	/**
-	 * Removes from an inventory slot (first arg) up to a specified number (second arg) of items and returns them in a
-	 * new stack.
+	/* (non-Javadoc)
+	 * @see net.minecraft.tileentity.TileEntity#hasCapability(net.minecraftforge.common.capabilities.Capability, net.minecraft.util.EnumFacing)
 	 */
 	@Override
-	public ItemStack decrStackSize(int slot, int number)
-	{
-		if (this.contents[slot] != null)
-		{
-			ItemStack itemstack;
-
-			if (this.contents[slot].stackSize <= number)
-			{
-				itemstack = this.contents[slot];
-				this.contents[slot] = null;
-				return itemstack;
-			}
-			else
-			{
-				itemstack = this.contents[slot].splitStack(number);
-
-				if (this.contents[slot].stackSize == 0)
-				{
-					this.contents[slot] = null;
-				}
-
-				return itemstack;
-			}
-		}
-		else
-		{
-			return null;
-		}
+	public boolean hasCapability(Capability<?> capability, EnumFacing facing) {
+		if(capability==CapabilityItemHandler.ITEM_HANDLER_CAPABILITY)return true;
+		return super.hasCapability(capability, facing);
 	}
-	/**
-	 * Sets the given item stack to the specified slot in the inventory (can be crafting or armor sections).
+	/* (non-Javadoc)
+	 * @see net.minecraft.tileentity.TileEntity#getCapability(net.minecraftforge.common.capabilities.Capability, net.minecraft.util.EnumFacing)
 	 */
 	@Override
-	public void setInventorySlotContents(int slot, ItemStack stack) {
-		contents[slot]=stack;
-		//if stacksize > InventoryStackLimit && Shift-clicking the stack in, the "extra" items are deleted...
-		if (stack != null && stack.stackSize > getInventoryStackLimit()) {
-			stack.stackSize = getInventoryStackLimit();
-		}              
-	}
-
-	@Override
-	/**
-	 * When some containers are closed they call this on each slot, then drop whatever it returns as an EntityItem -
-	 * like when you close a workbench GUI.
-	 * <br/>NOT to be used for the Dust Dye container
-	 */
-	public ItemStack removeStackFromSlot(int slot) {
-		ItemStack stack = getStackInSlot(slot);
-		if (stack != null) {
-			setInventorySlotContents(slot, null);
-		}
-		return stack;
-	}
-
-	@Override
-	/**
-	 * Returns the name of the inventory
-	 */
-	public String getName()
-	{
-		return References.modid+".DustDye";
-	}
-
-
-	@Override
-	public boolean isUseableByPlayer(EntityPlayer player) {
-		return worldObj.getTileEntity(pos) == this &&
-				player.getDistanceSq(pos.getX() + 0.5, pos.getY() + 0.5, pos.getZ() + 0.5) < 64;
-	}
-	//No clue what the next 2 methods 2 (NOT run when inv. opened)
-	@Override
-	public void openInventory(EntityPlayer p) {
-
-	}
-
-	@Override
-	public void closeInventory(EntityPlayer p) {
-	}
-
-	@Override
-	public boolean isItemValidForSlot(int slot, ItemStack stack) {
-		//only allow dyed dust in the dyer
-		return slot==0 ? stack.getItem() instanceof DustDyed : false;
-	}
-	@Override
-	public int getInventoryStackLimit() {
-		return 64;
-	}
-
-	@Override
-	public boolean hasCustomName() {
-		return true;
+	public <T> T getCapability(Capability<T> capability, EnumFacing facing) {
+		if(capability==CapabilityItemHandler.ITEM_HANDLER_CAPABILITY)return CapabilityItemHandler.ITEM_HANDLER_CAPABILITY.cast(inventory); 
+		return super.getCapability(capability, facing);
 	}
 	//might want to change the tag names to be variables for "safety" (nah, "safety" is overrated)
 	@Override
 	public void readFromNBT(NBTTagCompound tagCompound) {
 		super.readFromNBT(tagCompound);
-
-		NBTTagList tagList = tagCompound.getTagList("Inventory",10);
-		for (int i = 0; i < tagList.tagCount(); i++) {
-			NBTTagCompound tag = tagList.getCompoundTagAt(i);
-			byte slot = tag.getByte("Slot");
-			if (slot >= 0 && slot < contents.length) {
-				contents[slot] = ItemStack.loadItemStackFromNBT(tag);
-			}
-		}
+		NBTTagCompound contents = tagCompound.getCompoundTag("contents");
+		inventory = new DyedDustItemStackHandler(this);
+		inventory.deserializeNBT(contents);
+		//not syncing to the client on world load
 		this.colorString=tagCompound.getString("Color");
 		this.pastRedstoneState=tagCompound.getBoolean("pastRedstone");
 	}
@@ -193,52 +122,40 @@ public class TileEntityDustDye extends TileEntity implements IInventory{
 	@Override
 	public NBTTagCompound writeToNBT(NBTTagCompound tagCompound) {
 		super.writeToNBT(tagCompound);
-
-		NBTTagList itemList = new NBTTagList();
-		for (int i = 0; i < contents.length; i++) {
-			ItemStack stack = contents[i];
-			if (stack != null) {
-				NBTTagCompound tag = new NBTTagCompound();
-				tag.setByte("Slot", (byte) i);
-				stack.writeToNBT(tag);
-				itemList.appendTag(tag);
-			}
-		}
-		tagCompound.setTag("Inventory", itemList);
+		tagCompound.setTag("contents", inventory.serializeNBT());
 		tagCompound.setString("Color", colorString);
 		tagCompound.setBoolean("pastRedstone", pastRedstoneState);
 		return tagCompound;
 	}
-
-	@Override
-	public ITextComponent getDisplayName() {
-		//line from InventoryBasic
-		//might want to always return the translated version? (dosen't seem much used, leave as is.)
-		return this.hasCustomName() ? new TextComponentString(this.getName()) : new TextComponentTranslation(this.getName());
-	}
-	/*[refactor] the following methods are used to change the fields of the TileEntity.
-	 * we are already doing this in other ways, but we might want to switch to them eventually
-	 */
-	@Override
-	public int getField(int id) {
-		//not using this
-		return 0;
-	}
-	@Override
-	public void setField(int id, int value) {
-		//not using this?
-	}
-	@Override
-	public int getFieldCount() {
-		//Not using this?
-		return 0;
-	}
-	@Override
-	public void clear() {
-		// let's just do the same thing as inventoryBasic
-		for(int i=0;i<contents.length;i++){
-			contents[i]=null;
+	
+	private static class DyedDustItemStackHandler extends ItemStackHandler{
+		private final TileEntityDustDye te;
+		public DyedDustItemStackHandler(TileEntityDustDye tile) {
+			te = tile;
 		}
+		/* (non-Javadoc)
+		 * @see net.minecraftforge.items.ItemStackHandler#insertItem(int, net.minecraft.item.ItemStack, boolean)
+		 */
+		@Override
+		public ItemStack insertItem(int slot, ItemStack stack, boolean simulate) {
+			if(stack==null || stack.getItem()==WizardryRegistry.dust_dyed){
+				return super.insertItem(slot, stack, simulate);
+			}else if(stack.getItem() instanceof ItemDustPouch){
+				ItemDustPouch pouch = (ItemDustPouch)stack.getItem();
+				ItemStack dust = pouch.getDustStack(stack, 0);
+				if(dust!=null && dust.getItem()==WizardryRegistry.dust_dyed){
+					return super.insertItem(slot, stack, simulate);
+				}
+			}
+			return stack;
+		}
+		/* (non-Javadoc)
+		 * @see net.minecraftforge.items.ItemStackHandler#onContentsChanged(int)
+		 */
+		@Override
+		protected void onContentsChanged(int slot) {
+			te.markDirty();
+		}
+		
 	}
-
 }